[FLINK-36245] Relocate deprecated SourceFunction / SinkFunction and remove deprecated Sink V1 in 2.0 (#25331)
